Rename "compound-schema-types.ts" to "compounds.ts"
