+++ src/types/compounds.ts
